# docs/api/openapi-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Fragrance AI API",
    "description": "🌸 완벽한 AI 향수 추천 및 생성 플랫폼\n\n## 주요 기능\n- 🧠 AI 기반 향수 레시피 생성\n- 🔍 의미 기반 향수 검색\n- 📊 실시간 성능 모니터링\n- 🛡️ 엔터프라이즈 보안\n\n## 인증\nBearer 토큰을 사용한 JWT 인증이 필요합니다.\n\n```bash\nAuthorization: Bearer YOUR_JWT_TOKEN\n```",
    "version": "2.0.0",
    "contact": {
      "name": "Fragrance AI Team",
      "email": "api@fragrance-ai.com",
      "url": "https://fragrance-ai.com"
    },
    "license": {
      "name": "Proprietary License",
      "url": "https://fragrance-ai.com/license"
    },
    "termsOfService": "https://fragrance-ai.com/terms"
  },
  "servers": [
    {
      "url": "https://api.fragrance-ai.com/api/v2",
      "description": "Production Server"
    },
    {
      "url": "https://staging-api.fragrance-ai.com/api/v2",
      "description": "Staging Server"
    },
    {
      "url": "http://localhost:8000/api/v2",
      "description": "Development Server"
    }
  ],
  "paths": {
...
    "/generate/recipe": {
      "post": {
        "tags": ["Generation"],
        "summary": "🧪 향수 레시피 생성",
        "description": "AI를 사용하여 맞춤형 향수 레시피를 생성합니다.",
        "operationId": "generate_recipe",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeGenerationRequest"
              },
              "examples": {
                "romantic_perfume": {
                  "summary": "로맨틱 향수",
                  "value": {
                    "mood": "romantic",
                    "season": "spring",
                    "intensity": "moderate",
                    "notes_preference": ["rose", "jasmine", "vanilla"],
                    "target_audience": "women",
                    "occasion": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "레시피 생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeGenerationResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RecipeGenerationRequest": {
        "type": "object",
        "required": ["mood"],
        "properties": {
          "mood": {
            "type": "string",
            "description": "원하는 무드",
            "example": "romantic"
          },
          "season": {
            "type": "string",
            "enum": ["spring", "summer", "autumn", "winter"],
            "description": "계절"
          },
          "intensity": {
            "type": "string",
            "enum": ["light", "moderate", "strong"],
            "default": "moderate"
          },
          "notes_preference": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "선호하는 향료"
          },
          "target_audience": {
            "type": "string",
            "enum": ["men", "women", "unisex"],
            "default": "unisex"
          },
          "occasion": {
            "type": "string",
            "description": "사용 상황",
            "example": "daily"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": true
          },
          "error_code": {
            "type": "string",
            "example": "E0001"
          },
          "message": {
            "type": "string",
            "example": "Invalid request parameter"
          },
          "details": {
            "type": "object"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "잘못된 요청",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "인증 실패",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT 토큰을 사용한 인증"
      }
    }
  },
...
}